Fix remaining issues with tests
