+++ kifi-backend/modules/shoebox/app/com/keepit/notify/ShoeboxNotifyModule.scala
@@ -9,3 +9,4 @@ case class ShoeboxNotifyModule() extends NotifyModule {
   }
 
 }
+
